bundle dragonruby with the repository
according to dragonruby's docs, this is the way to ensure that your game
doesn't break in the long term. makes sense, but i wanted to make it
easier to be able to tell exactly which version my dragonruby binary is
at a glance, and at what points i upgraded it, etc.

so i'm including the SHA in the binary's filename. i retrieved the SHA
when running the binary, from the server logs in my terminal.