refactor `s/Level#execute!/Level#tick!`
it seems to make sense to just call this `#tick!` to communicate that it
needs to be called within the global `#tick` method to re-render with
each frame.

while i was renaming, i was able to devise a nice way to test the
functionality of levels (there isn't much functionality, but it's
important).